Share directory scanning between node & browser
