+++ src/plugins/plugin-index.ts
@@ -1,11 +1,9 @@
 // # file-index.js
 import os from 'node:os';
-import { DBPF, FileType } from 'sc4/core';
 import { TGIIndex } from 'sc4/utils';
 import FileScanner from './file-scanner.js';
-import type { Entry } from 'sc4/core';
-import createLoadComparator from './create-load-comparator.js';
 import CorePluginIndex from './core-plugin-index.js';
+import DirectoryScanOperation from './directory-scan-operation.js';
 
 type folder = string;
 type PluginIndexOptions = {
@@ -14,7 +12,6 @@ type PluginIndexOptions = {
 	installation?: folder;
 	plugins?: folder;
 	mem?: number;
-	threads?: number;
 };
 
 type BuildOptions = {
@@ -34,7 +31,6 @@ export default class PluginIndex extends CorePluginIndex {
 		core: boolean;
 		installation: folder | undefined;
 		plugins: folder | undefined;
-		threads: number | undefined;
 	};
 
 	// ## constructor(opts)
@@ -58,13 +54,11 @@ export default class PluginIndex extends CorePluginIndex {
 			core = true,
 			installation = process.env.SC4_INSTALLATION,
 			plugins = process.env.SC4_PLUGINS,
-			threads,
 		} = opts;
 		this.options = {
 			core,
 			installation,
 			plugins,
-			threads,
 		};
 
 	}
@@ -82,14 +76,16 @@ export default class PluginIndex extends CorePluginIndex {
 			installation = this.options.installation,
 		} = opts;
 		if (this.options.core && installation) {
-			let op = new DirectoryScanOperation(this, installation);
+			const glob = new FileScanner(this.scan, { cwd: installation });
+			const op = new DirectoryScanOperation(this, glob);
 			all.push(op.start());
 			ops.push(op);
 		}
 
 		// Same for the user plugins folder.
 		if (plugins) {
-			let op = new DirectoryScanOperation(this, plugins);
+			const glob = new FileScanner(this.scan, { cwd: plugins });
+			const op = new DirectoryScanOperation(this, glob);
 			all.push(op.start());
 			ops.push(op);
 		}
@@ -108,72 +104,4 @@ export default class PluginIndex extends CorePluginIndex {
 		return this;
 	}
 
-}
-
-type QueueItem = {
-	dbpf: DBPF;
-	entries: Entry[];
-};
-
-class DirectoryScanOperation {
-	index: PluginIndex;
-	path: string;
-	dbpfs: DBPF[] = [];
-	queue: QueueItem[] = [];
-	filesPromise = Promise.withResolvers<DBPF[]>();
-	entriesPromise = Promise.withResolvers<Entry[]>();
-	constructor(index: PluginIndex, path: string) {
-		this.index = index;
-		this.path = path;
-	}
-	async start() {
-
-		// First we'll look up all the dbpf files in this folder. Once we have 
-		// the files, we can already resolve this promise so that a progress 
-		// indicator might show how many files we have found in total, while 
-		// being non-blocking to wait for it.
-		const glob = new FileScanner(this.index.scan, { cwd: this.path });
-		const tasks: Promise<any>[] = [];
-		for await (let file of glob) {
-			const dbpf = new DBPF(file);
-			this.dbpfs.push(dbpf);
-
-			// Immediately start reading in the DBPF.
-			const item: QueueItem = { dbpf, entries: [] };
-			this.queue.push(item);
-			const task = dbpf.parseAsync().then(() => {
-				for (let entry of dbpf) {
-					if (entry.type !== FileType.DIR) {
-						item.entries.push(entry);
-					}
-				}
-			});
-			tasks.push(task);
-
-		}
-		this.filesPromise.resolve(this.dbpfs);
-
-		// If we reach this point, all dbpfs have been added to the queue and 
-		// they have started loading. This means that we can sort them now based 
-		// on the load order that is required!
-		const compare = createLoadComparator();
-		this.queue.sort((a, b) => compare(a.dbpf.filename, b.dbpf.filename));
-
-		// Cool, now wait for the entries to be finished parsing as well. Once 
-		// that's done, the queue is properly sorted and contain all entries. 
-		// The last thing we have to do hence is flatten the entries. Note that 
-		// internally, DBPF's don't use LIFO, but FIFO, meaning that TGI's that 
-		// appear *first* in the dbpf actually get preference, so we need to 
-		// take this into account!
-		await Promise.all(tasks);
-		const flat = [];
-		for (let { entries } of this.queue) {
-			for (let i = entries.length-1; i >= 0; i--) {
-				flat.push(entries[i]);
-			}
-		}
-		this.entriesPromise.resolve(flat);
-		return flat;
-
-	}
 }
